chore: clean up syntax and add tests
